Refactor icon rendering to be used by both title/description in timetable details, implement updating description & beginning of designation display, refactor course list in explore page to call separate component for each course card
