+++ frontend/src/components/Timetable.jsx
@@ -9,6 +9,7 @@ import {
   TITLE_ERROR_MESSAGE,
   GENERIC_ERROR,
   TITLE_PATH,
+  DESCRIPTION_PATH,
 } from "../utils/constants";
 
 const Timetable = () => {
@@ -18,6 +19,52 @@ const Timetable = () => {
   const [isEditingTitle, setIsEditingTitle] = useState(false);
   const [title, setTitle] = useState("");
   const [titleError, setTitleError] = useState();
+  const [isEditingDescription, setIsEditingDescription] = useState(false);
+  const [description, setDescription] = useState("");
+  const [descriptionError, setDescriptionError] = useState("");
+  const TIMETABLE_DESCRIPTION = "Timetable Description ";
+  const DESIGNATION = "Designation";
+
+  const cancelEditingDescription = () => {
+    setDescription(timetable?.description);
+    setIsEditingDescription(false);
+    setDescriptionError("");
+  };
+
+  const updateDescription = async () => {
+    setIsEditingDescription(false);
+    setDescriptionError("");
+
+    if (description === timetable?.description) {
+      return;
+    }
+
+    try {
+      const response = await fetch(
+        `${import.meta.env.VITE_BASE_URL}${
+          Path.TIMETABLE
+        }${DESCRIPTION_PATH}${ID_QUERY_PARAM}${timetable?.id}`,
+        {
+          method: "PATCH",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({ description }),
+          credentials: "include",
+        }
+      );
+
+      if (response.ok) {
+        fetchTimetableData(timetable?.id);
+      } else {
+        setDescription(timetable?.description);
+        setDescriptionError(GENERIC_ERROR);
+      }
+    } catch (error) {
+      setDescription(timetable?.description);
+      setDescriptionError(GENERIC_ERROR);
+    }
+  };
 
   const cancelEditingTitle = () => {
     setTitle(timetable?.title);
@@ -47,7 +94,7 @@ const Timetable = () => {
           headers: {
             "Content-Type": "application/json",
           },
-          body: JSON.stringify({title}),
+          body: JSON.stringify({ title }),
           credentials: "include",
         }
       );
@@ -79,6 +126,7 @@ const Timetable = () => {
       if (response.ok) {
         const data = await response.json();
         setTitle(data?.timetable?.title);
+        setDescription(data?.timetable?.description);
         setTimetable(data?.timetable);
       } else {
         navigate(Path.EXPLORE);
@@ -86,6 +134,36 @@ const Timetable = () => {
     } catch (error) {
       navigate(Path.EXPLORE);
     }
+  };
+
+  const renderIcons = (isEditing, setIsEditing, updateItem, cancelEditing) => {
+    return (
+      <>
+        {isEditing ? (
+          <>
+            <span
+              className="material-symbols-outlined timetable-edit-icon timetable-yes"
+              onClick={updateItem}
+            >
+              check_circle
+            </span>
+            <span
+              className="material-symbols-outlined timetable-edit-icon timetable-no"
+              onClick={cancelEditing}
+            >
+              cancel
+            </span>
+          </>
+        ) : (
+          <span
+            className="material-symbols-outlined timetable-edit-icon"
+            onClick={() => setIsEditing(true)}
+          >
+            edit
+          </span>
+        )}
+      </>
+    );
   };
 
   useEffect(() => {
@@ -108,6 +186,7 @@ const Timetable = () => {
           west
         </span>
       </button>
+
       <div className="text-input-container timetable-title">
         <div className="timetable-edit-container">
           <input
@@ -119,33 +198,51 @@ const Timetable = () => {
             onChange={(event) => setTitle(event.target.value)}
             disabled={!isEditingTitle}
           />
-          {isEditingTitle ? (
-            <>
-              <span
-                className="material-symbols-outlined timetable-edit-icon timetable-yes"
-                onClick={updateTitle}
-              >
-                check_circle
-              </span>
-              <span
-                className="material-symbols-outlined timetable-edit-icon timetable-no"
-                onClick={cancelEditingTitle}
-              >
-                cancel
-              </span>
-            </>
-          ) : (
-            <span
-              className="material-symbols-outlined timetable-edit-icon"
-              onClick={() => setIsEditingTitle(true)}
-            >
-              edit
-            </span>
+          {renderIcons(
+            isEditingTitle,
+            setIsEditingTitle,
+            updateTitle,
+            cancelEditingTitle
           )}
         </div>
         <div className="create-email-section-spacing text-input-error dropdown-input-error timetable-input-error">
           {titleError}
         </div>
+      </div>
+
+      <label
+        htmlFor="description"
+        className="page-big-header timetable-description"
+      >
+        {TIMETABLE_DESCRIPTION}
+        <span className="timetable-icon-container">
+          {renderIcons(
+            isEditingDescription,
+            setIsEditingDescription,
+            updateDescription,
+            cancelEditingDescription
+          )}
+        </span>
+      </label>
+      <div className="text-input-container timetable-description">
+        <textarea
+          id="description"
+          className="text-input create-email-body timetable-edit-disabled"
+          placeholder={TIMETABLE_DESCRIPTION_PLACEHOLDER}
+          value={description}
+          onChange={(event) => setDescription(event.target.value)}
+          disabled={!isEditingDescription}
+        />
+        <div className="create-email-section-spacing text-input-error dropdown-input-error">
+          {descriptionError}
+        </div>
+      </div>
+
+      <div htmlFor="designation" className="page-big-header timetable-description">
+        {DESIGNATION}
+      </div>
+      <div className="text-input-container">
+        {}
       </div>
     </div>
   );
