+++ frontend/src/utils/constants.js
@@ -401,7 +401,7 @@ export const CC = "CC";
 export const CART_PATH = "/cart";
 export const FAVORITES_PATH = "/favorites";
 export const TITLE_PATH = "/title";
-export const DESCRIPTION_PATH = "/title";
+export const DESCRIPTION_PATH = "/description";
 export const ID_QUERY_PARAM = "?id=";
 
 export const YES = "Yes";
